+++ next-docs/components/codePreview/Preview.tsx
@@ -24,6 +24,10 @@ const Preview = ({ title, preview, code, isGrayBg }: PreviewProps) => {
   const [isPreviewActive, setActive] = useState(true);
   const setPreviewActive = () => setActive(true);
   const setCodeActive = () => setActive(false);
+  const copyCode = () => {
+    if (navigator?.clipboard) navigator.clipboard.writeText(code ? code : "");
+  };
+
   return (
     <>
       <div className="flex flex-row items-center justify-between">
@@ -34,6 +38,7 @@ const Preview = ({ title, preview, code, isGrayBg }: PreviewProps) => {
           isPreviewActive={isPreviewActive}
           setPreviewActive={setPreviewActive}
           setCodeActive={setCodeActive}
+          copyCode={copyCode}
         />
       </div>
       <div