+++ next-docs/public/examples/sidebar/Rtl.tsx
@@ -1,21 +1,22 @@
 import React from 'react';
 import { Sidebar } from '@heathmont/moon-sidebar';
 import {
-  IconChevronLeftRounded,
-  IconChevronRightRounded,
-  IconBitcoinTalk,
-} from '@heathmont/moon-assets';
+  ControlsChevronLeft,
+  ControlsChevronRight,
+  ShopCryptoCoin,
+} from '@heathmont/moon-icons';
 import { Inline } from '@heathmont/moon-components';
+import { rem } from "@heathmont/moon-utils";
 
 const Example = () => {
   return (
     <>
       <Sidebar
-        collapseIcon={<IconChevronRightRounded />}
-        expandIcon={<IconChevronLeftRounded />}
+        collapseIcon={<ControlsChevronLeft fontSize={rem(23)}/>}
+        expandIcon={<ControlsChevronRight fontSize={rem(23)}/>}
         logo={
           <Inline fontSize="2rem">
-            <IconBitcoinTalk />
+            <ShopCryptoCoin color='krillin.100'/>
           </Inline>
         }
         isOpen={false}