+++ packages/core/src/tabLink/TabLink.tsx
@@ -13,10 +13,19 @@ export interface TabLinkProps {
 
 const ElementLeftWrapper = styled.div({
   marginInlineEnd: rem(8),
+  '& > svg': {
+    width: rem(15),
+    height: rem(15),
+    margin: rem(4.5)
+  }
 });
 
 const CountWrapper = styled.span({
   marginInlineStart: rem(8),
+  width: rem(24),
+  display: 'flex',
+  alignItems: 'center',
+  justifyContent: 'center'
 });
 
 const TabLink: React.FC<TabLinkProps> = ({
@@ -31,7 +40,6 @@ const TabLink: React.FC<TabLinkProps> = ({
     size={size}
     isTop={isTop}
     href={href}
-    paddingSize={size === 'small' ? rem(2) : rem(6)}
   >
     {elementLeft && <ElementLeftWrapper>{elementLeft}</ElementLeftWrapper>}
     {children}